recommendation/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymysql>=1.1.1
cryptography>=43.0.0
psycopg2-binary>=2.9.9
pgvector>=0.3.6
numpy>=1.26.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0

pydantic